add tests

+++ core/domain/src/main/kotlin/fr/r4phab/weather/domain/models/WeatherForecastModel.kt
@@ -6,6 +6,6 @@ data class WeatherForecastModel(
     val day: DayOfWeek,
     val weather: WeatherModel,
     val dayNight: DayNightModel,
-    val minimumTemperature: Float,
-    val maximumTemperature: Float,
+    val minimumTemperature: Double,
+    val maximumTemperature: Double,
 )
